requirements.txt
streamlit
uuid
gcsfs 
google-cloud-storage
pillow 
numpy
tensorflow